+++ kafka-junit4/src/test/java/com/salesforce/kafka/test/junit4/SharedKafkaTestResourceWithSaslPlainTest.java
@@ -26,39 +26,16 @@
 package com.salesforce.kafka.test.junit4;
 
 import com.salesforce.kafka.test.KafkaTestUtils;
-import com.salesforce.kafka.test.listeners.SslListener;
-import org.apache.kafka.clients.consumer.ConsumerRecord;
-import org.apache.kafka.clients.consumer.ConsumerRecords;
-import org.apache.kafka.clients.consumer.KafkaConsumer;
-import org.apache.kafka.clients.producer.KafkaProducer;
-import org.apache.kafka.clients.producer.ProducerRecord;
-import org.apache.kafka.clients.producer.RecordMetadata;
-import org.apache.kafka.common.Node;
-import org.apache.kafka.common.PartitionInfo;
-import org.apache.kafka.common.TopicPartition;
-import org.apache.kafka.common.serialization.StringDeserializer;
-import org.apache.kafka.common.serialization.StringSerializer;
+import com.salesforce.kafka.test.listeners.SaslPlainListener;
+
 import org.junit.ClassRule;
-import org.junit.Test;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-
-import java.util.ArrayList;
-import java.util.Collection;
-import java.util.List;
-import java.util.Set;
-import java.util.concurrent.Future;
-import java.util.stream.Collectors;
-
-import static org.junit.Assert.assertEquals;
-import static org.junit.Assert.assertNotNull;
-import static org.junit.Assert.assertTrue;
 
 /**
- * Run smoke tests against an SSL enabled cluster.
- * @see AbstractSharedKafkaTestResourceTest for test case definitions.
+ * NOTE: This test case assumes you've started the JVM with the argument.
+ *
+ * -Djava.security.auth.login.config=kafka-junit-core/src/test/resources/jaas.conf
  */
-public class SharedKafkaTestResourceWithSslTest extends AbstractSharedKafkaTestResourceTest {
+public class SharedKafkaTestResourceWithSaslPlainTest extends AbstractSharedKafkaTestResourceTest {
     /**
      * We have a single embedded kafka server that gets started when this test class is initialized.
      *
@@ -74,14 +51,10 @@ public class SharedKafkaTestResourceWithSslTest extends AbstractSharedKafkaTestR
         .withBrokers(2)
         // Disable topic auto-creation.
         .withBrokerProperty("auto.create.topics.enable", "false")
-        // Register and configure SSL authentication on cluster.
-        .registerListener(new SslListener()
-        .useSslForInterBrokerProtocol()
-        .withKeyStoreLocation(SharedKafkaTestResourceWithSslTest.class.getClassLoader().getResource("kafka.keystore.jks").getFile())
-        .withKeyStorePassword("password")
-        .withTrustStoreLocation(SharedKafkaTestResourceWithSslTest.class.getClassLoader().getResource("kafka.truststore.jks").getFile())
-        .withTrustStorePassword("password")
-        .withKeyPassword("password")
+        // Register and configure SASL PLAIN authentication on cluster.
+        .registerListener(new SaslPlainListener()
+        .withUsername("kafkaclient")
+        .withPassword("client-secret")
     );
 
     /**